refactor column/template behaviors
